Add behavior to delete or cancel a row

+++ src/resources/removeModels.js
@@ -1,17 +1,30 @@
 import {deleteData} from './Requests';
 import {routeMap} from '../routes/RouteMap';
 
-const removeClient = (items, key) => removeFromList(routeMap.dependent, items, key);
+const removeClient = (items, key, remove) => removeFromList(routeMap.dependent, items, key, remove);
 
-const removeRegularCost = (items, key) => removeFromList(routeMap.regular_cost, items, key);
+const removeRegularCost = (items, key, remove) => removeFromList(routeMap.regular_cost, items, key, remove);
 
-const removeGoal = (items, key) => removeFromList(routeMap.goal, items, key);
+const removeGoal = (items, key, remove) => removeFromList(routeMap.goal, items, key, remove);
 
-const removePatrimony = (patrimony, items, key) => removeFromList(routeMap[patrimony], items, key);
-const removeFromList = (route, items, key) => {
+const removePatrimony = (patrimony, items, key, remove) => removeFromList(routeMap[patrimony], items, key, remove);
+
+// Remove or only cancel the edit in row
+const removeFromList = (route, items, key, remove) => {
   const item = items.find( item => item.index === key);
-  deleteData(`${route}${item.id}/`);
-  return items.filter( item => item.index !== key);
+  if (remove) { // only items with ID can be removed
+    deleteData(`${route}${item.id}/`);
+    return items.filter( item => item.index !== key);
+  } else if (item.id) { // back item to previous state
+    return items.map( item => {
+      if (item.index === key) {
+        item.selected = false;
+      }
+      return item;
+    });
+  } else { // is a row not persisted, only remove it
+    return items.filter( item => item.index !== key);
+  }
 };
 
 export {removeClient, removeRegularCost, removeGoal, removePatrimony};
